refactor: Refactor box layer components into one folder
