+++ ionic_user_interface/src/components/wall-image-viewer/box-layer/boundingBoxAttributes.ts
