+++ ionic_user_interface/src/components/wall-image-viewer/box-layer/index.ts
@@ -1,0 +1,3 @@
+import { useBoxLayer } from './useBoxLayer';
+
+export { useBoxLayer };
