--- ionic_user_interface/src/components/wall-image-viewer/enums.ts
@@ -1,35 +1,0 @@
-enum BoxClass {
-  HOLD,
-  VOLUME,
-  DRAWN,
-}
-
-enum BoxState {
-  NORMAL_HANDHOLD,
-  FOOTHOLD,
-  SINGLE_START_HANDHOLD,
-  DUAL_START_HANDHOLD,
-  END_HANDHOLD,
-  UNSELECTED,
-  HIDDEN,
-}
-
-enum TapeMode {
-  SINGLE_START,
-  DUAL_START,
-  NONE,
-}
-
-enum NumberMode {
-  ON,
-  OFF,
-}
-
-enum SelectMode {
-  HANDHOLD,
-  FOOTHOLD,
-  DRAWBOX,
-  MARKDONE,
-}
-
-export { BoxClass, BoxState, SelectMode, TapeMode, NumberMode };
